--- stories/CrossOrigin/crossOrigin.css
@@ -1,4 +1,0 @@
-button {
-    display: block;
-    margin: 10px;
-}
