+++ stories/Observed/Observed.ts
@@ -5,7 +5,10 @@
 
 import "./observed.css";
 import {
+    createTabster,
+    getCrossOrigin,
     getCurrentTabster,
+    getObservedElement,
     getTabsterAttribute,
     Types as TabsterTypes,
 } from "tabster";
@@ -14,12 +17,12 @@ export type ObservedElementProps = TabsterTypes.ObservedElementProps;
 
 const DELAY = 1000;
 
-export const createAsyncObservedWrapper = (props: ObservedElementProps) => {
+export const createObservedWrapper = (props: ObservedElementProps) => {
     const { name } = props;
 
     // create observed target
     const observedContainer = document.createElement("div");
-    const observedTarget = createObserved(props);
+    const observedTarget = createObserved(props, document);
     const mountObservedTargetWithDelay = () => {
         if (observedContainer.childElementCount) {
             observedContainer.removeChild(observedTarget);
@@ -30,58 +33,112 @@ export const createAsyncObservedWrapper = (props: ObservedElementProps) => {
     };
 
     // create multiple triggers buttons
-    const trigger = createTrigger({
-        name,
-        text: `Asynchronously show and focus observed element with name ${name}`,
-        onClick: mountObservedTargetWithDelay,
-    });
+    const triggers = [name].map((name) =>
+        createTrigger({
+            name,
+            innerText: `Asynchronously show and focus observed element with name ${name}`,
+            onClick: mountObservedTargetWithDelay,
+        })
+    );
 
     const wrapper = document.createElement("div");
-    wrapper.appendChild(trigger);
+    triggers.forEach((trigger) => wrapper.appendChild(trigger));
     wrapper.appendChild(observedContainer);
 
     return wrapper;
 };
 
-export const createObservedWrapper = (props: ObservedElementProps) => {
+const setupTabsterInIframe = (currWindow: Window) => {
+    const tabster = createTabster(currWindow, {
+        autoRoot: {},
+        controlTab: true,
+        rootDummyInputs: undefined,
+    });
+    console.log("created tabster for iframe");
+    getObservedElement(tabster);
+    getCrossOrigin(tabster);
+    tabster?.crossOrigin?.setup();
+    console.log("created cross origin");
+};
+
+declare global {
+    interface Window {
+        setupTabsterInIframe: (currWindow: Window) => void;
+    }
+}
+window.setupTabsterInIframe = setupTabsterInIframe;
+
+export const createObservedWrapperWithIframe = (
+    props: ObservedElementProps
+) => {
     const { name } = props;
 
-    // create observed target
-    const observedTarget = createObserved(props);
+    const iframe = document.createElement("iframe");
+    // Note: dynamic iframe using srcdoc does not work https://bugs.chromium.org/p/chromium/issues/detail?id=1339813
+    // a `src` attribute is required
+    iframe.src = `#`;
+    iframe.onload = () => {
+        const document = iframe.contentDocument;
+        if (document) {
+            document.body.innerHTML = "";
+            const script = document.createElement("script");
+            script.innerText = `setupTabsterInIframe(window)`;
+            document.body.append(script);
+            const observedTarget = createObserved(props, document);
+            document.body.append(observedTarget);
+        }
+    };
+    iframe.width = "600px";
 
-    // create multiple triggers buttons
-    const trigger = createTrigger({
-        name,
-        text: `Focus observed element with name ${name}`,
-    });
+    // create trigger button
+    const triggers = [name].map((name) =>
+        createTrigger({
+            name,
+            innerText: `Focus observed element in iframe with name ${name}`,
+            isCrossOrigin: true,
+        })
+    );
 
     const wrapper = document.createElement("div");
-    wrapper.appendChild(trigger);
-    wrapper.appendChild(observedTarget);
+    triggers.forEach((trigger) => wrapper.appendChild(trigger));
+    wrapper.appendChild(iframe);
 
     return wrapper;
 };
 
 type TriggerProps = {
     name: string;
-    text: string;
+    innerText: string;
     onClick?: () => void;
+    isCrossOrigin?: boolean;
 };
-const createTrigger = ({ name, text, onClick }: TriggerProps) => {
+const createTrigger = ({
+    name,
+    innerText,
+    onClick,
+    isCrossOrigin,
+}: TriggerProps) => {
     const trigger = document.createElement("button");
     trigger.id = `trigger-for-${name}`;
-    trigger.innerText = text;
+    trigger.innerText = innerText;
     trigger.onclick = function () {
         onClick?.();
         const tabster = getCurrentTabster(window);
-        tabster?.observedElement?.requestFocus(name, 5000);
+        if (isCrossOrigin) {
+            tabster?.crossOrigin?.observedElement?.requestFocus(name, 5000);
+        } else {
+            tabster?.observedElement?.requestFocus(name, 5000);
+        }
     };
 
     return trigger;
 };
 
-const createObserved = (props: ObservedElementProps) => {
-    const observed = document.createElement("div");
+const createObserved = (
+    props: ObservedElementProps,
+    currDocument: Document
+) => {
+    const observed = currDocument.createElement("div");
     observed.tabIndex = 0;
     observed.classList.add("observed");
     observed.innerText = `observed element with name: ${JSON.stringify(
